+++ src/index.ts
@@ -2,4 +2,4 @@ import OurPlugin from "./OurPlugin";
 
 // Expose our class as a PostCSS Plugin.
 const ourPlugin = OurPlugin.asPostCSSPlugin();
-export default ourPlugin;
+export = ourPlugin;