+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 Releases ordered so that the most recent are displayed at the top, with the currently being developed release at the top, labeled as **In Development**. This release will be given a number once it is ready to be released. Each release can contain both a **Features and Improvements** and **Bug Fixes** sections.
 
+## In Development
+
+### Features and Improvements
+
+-   Add integration tests to validate plugin defined correctly in build.
+
+### Bug Fixes
+
+-   Fix error in main file definition.
+
 ## [1.8.0](https://github.com/dbtedman/postcss-prefixwrap/releases/tag/1.8.0)
 
 ### Features and Improvements